tfrecord works fine
